+++ src/constants/index.ts
@@ -1,5 +1,6 @@
 export * from "./env.constant";
 export * from "./error-types.constant";
+export * from "./events.constant";
 export * from "./magic-numbers.constant";
 export * from "./rbac.constants";
 export * from "./service.constant";